+++ packages/twenty-front/src/modules/debug/components/RecoilDebugObserver.tsx
@@ -1,6 +1,5 @@
-import { useRecoilTransactionObserver_UNSTABLE, useRecoilValue } from 'recoil';
+import { useRecoilTransactionObserver_UNSTABLE } from 'recoil';
 
-import { isDebugModeState } from '@/client-config/states/isDebugModeState';
 import { logDebug } from '~/utils/logDebug';
 
 const formatTitle = (stateName: string) => {
@@ -15,7 +14,7 @@ const formatTitle = (stateName: string) => {
 };
 
 export const RecoilDebugObserverEffect = () => {
-  const isDebugMode = useRecoilValue(isDebugModeState);
+  const isDebugMode = false;
 
   useRecoilTransactionObserver_UNSTABLE(({ snapshot }) => {
     if (!isDebugMode) {
