+++ packages/twenty-front/src/modules/object-record/record-field/hooks/useIsFieldEmpty.ts
@@ -9,6 +9,7 @@ import { FieldContext } from '../contexts/FieldContext';
 export const useIsFieldEmpty = () => {
   const { entityId, fieldDefinition, overridenIsFieldEmpty } =
     useContext(FieldContext);
+
   const fieldValue = useRecordFieldValue(
     entityId,
     fieldDefinition?.metadata?.fieldName ?? '',
