+++ packages/twenty-front/src/modules/object-record/record-table/components/RecordTableCellCheckbox.tsx
@@ -1,20 +1,15 @@
 import { useContext } from 'react';
-import styled from '@emotion/styled';
 
 import { CheckboxCell } from '@/object-record/record-table/components/CheckboxCell';
-import { StyledRecordTableTd } from '@/object-record/record-table/components/StyledRecordTableTd';
+import { RecordTableTd } from '@/object-record/record-table/components/RecordTableTd';
 import { RecordTableRowContext } from '@/object-record/record-table/contexts/RecordTableRowContext';
-
-const StyledCheckboxTd = styled(StyledRecordTableTd)`
-  border-right: transparent;
-`;
 
 export const RecordTableCellCheckbox = () => {
   const { isSelected, isDragging } = useContext(RecordTableRowContext);
 
   return (
-    <StyledCheckboxTd isSelected={isSelected}>
+    <RecordTableTd isSelected={isSelected}>
       {!isDragging && <CheckboxCell />}
-    </StyledCheckboxTd>
+    </RecordTableTd>
   );
 };
