+++ packages/twenty-front/src/modules/object-record/record-table/components/RecordTableCellsEmpty.tsx
@@ -1,7 +1,7 @@
 import { useContext } from 'react';
 import { useRecoilValue } from 'recoil';
 
-import { StyledRecordTableTd } from '@/object-record/record-table/components/StyledRecordTableTd';
+import { RecordTableTd } from '@/object-record/record-table/components/RecordTableTd';
 import { RecordTableRowContext } from '@/object-record/record-table/contexts/RecordTableRowContext';
 import { useRecordTableStates } from '@/object-record/record-table/hooks/internal/useRecordTableStates';
 
@@ -12,9 +12,6 @@ export const RecordTableCellsEmpty = () => {
   const visibleTableColumns = useRecoilValue(visibleTableColumnsSelector());
 
   return visibleTableColumns.map((column) => (
-    <StyledRecordTableTd
-      isSelected={isSelected}
-      key={column.fieldMetadataId}
-    ></StyledRecordTableTd>
+    <RecordTableTd isSelected={isSelected} key={column.fieldMetadataId} />
   ));
 };
