+++ packages/twenty-front/src/modules/object-record/record-table/components/RecordTableCellsVisible.tsx
@@ -1,11 +1,11 @@
 import { useContext } from 'react';
 import { useRecoilValue } from 'recoil';
 
-import { RecordTableCellFieldContextWrapper } from '@/object-record/record-table/components/RecordTableCellFieldContextWrapper';
-import { RecordTableCellContext } from '@/object-record/record-table/contexts/RecordTableCellContext';
 import { RecordTableRowContext } from '@/object-record/record-table/contexts/RecordTableRowContext';
 import { useRecordTableStates } from '@/object-record/record-table/hooks/internal/useRecordTableStates';
 import { RecordTableCell } from '@/object-record/record-table/record-table-cell/components/RecordTableCell';
+import { RecordTableCellTd } from '@/object-record/record-table/record-table-cell/components/RecordTableCellTd';
+import { RecordTableCellWrapper } from '@/object-record/record-table/record-table-cell/components/RecordTableCellWrapper';
 
 export const RecordTableCellsVisible = () => {
   const { isDragging } = useContext(RecordTableRowContext);
@@ -13,19 +13,27 @@ export const RecordTableCellsVisible = () => {
 
   const visibleTableColumns = useRecoilValue(visibleTableColumnsSelector());
 
-  return visibleTableColumns.map((column, columnIndex) => (
-    <RecordTableCellContext.Provider
-      value={{
-        columnDefinition: column,
-        columnIndex,
-      }}
-      key={column.fieldMetadataId}
-    >
-      {isDragging && columnIndex > 0 ? null : (
-        <RecordTableCellFieldContextWrapper>
+  const tableColumnsAfterFirst = visibleTableColumns.slice(1);
+
+  return (
+    <>
+      <RecordTableCellWrapper column={visibleTableColumns[0]} columnIndex={0}>
+        <RecordTableCellTd firstColumn>
           <RecordTableCell />
-        </RecordTableCellFieldContextWrapper>
-      )}
-    </RecordTableCellContext.Provider>
-  ));
+        </RecordTableCellTd>
+      </RecordTableCellWrapper>
+      {!isDragging &&
+        tableColumnsAfterFirst.map((column, columnIndex) => (
+          <RecordTableCellWrapper
+            key={column.fieldMetadataId}
+            column={column}
+            columnIndex={columnIndex + 1}
+          >
+            <RecordTableCellTd>
+              <RecordTableCell />
+            </RecordTableCellTd>
+          </RecordTableCellWrapper>
+        ))}
+    </>
+  );
 };
