+++ packages/twenty-front/src/modules/object-record/record-table/components/RecordTableLastEmptyCell.tsx
@@ -1,10 +1,10 @@
 import { useContext } from 'react';
 
-import { StyledRecordTableTd } from '@/object-record/record-table/components/StyledRecordTableTd';
+import { RecordTableTd } from '@/object-record/record-table/components/RecordTableTd';
 import { RecordTableRowContext } from '@/object-record/record-table/contexts/RecordTableRowContext';
 
 export const RecordTableLastEmptyCell = () => {
   const { isSelected } = useContext(RecordTableRowContext);
 
-  return <StyledRecordTableTd isSelected={isSelected} />;
+  return <RecordTableTd isSelected={isSelected} />;
 };
