+++ packages/twenty-front/src/modules/object-record/record-table/record-table-cell/components/RecordTableCellLoading.tsx
@@ -1,10 +1,10 @@
-import { StyledTd } from '@/object-record/record-table/components/RecordTableRow';
+import { RecordTableTd } from '@/object-record/record-table/components/RecordTableTd';
 import { RecordTableCellSkeletonLoader } from '@/object-record/record-table/record-table-cell/components/RecordTableCellSkeletonLoader';
 
 export const RecordTableCellLoading = () => {
   return (
-    <StyledTd>
+    <RecordTableTd>
       <RecordTableCellSkeletonLoader />
-    </StyledTd>
+    </RecordTableTd>
   );
 };
