+++ packages/twenty-front/src/modules/object-record/record-table/record-table-cell/hooks/useCurrentCellPosition.ts
@@ -1,21 +1,9 @@
-import { useContext, useMemo } from 'react';
+import { useContext } from 'react';
 
 import { RecordTableCellContext } from '@/object-record/record-table/contexts/RecordTableCellContext';
-import { RecordTableRowContext } from '@/object-record/record-table/contexts/RecordTableRowContext';
-
-import { TableCellPosition } from '../../types/TableCellPosition';
 
 export const useCurrentTableCellPosition = () => {
-  const { rowIndex } = useContext(RecordTableRowContext);
-  const { columnIndex } = useContext(RecordTableCellContext);
+  const { cellPosition } = useContext(RecordTableCellContext);
 
-  const currentTableCellPosition: TableCellPosition = useMemo(
-    () => ({
-      column: columnIndex,
-      row: rowIndex,
-    }),
-    [columnIndex, rowIndex],
-  );
-
-  return currentTableCellPosition;
+  return cellPosition;
 };
